refactor: move arquivos para Application\Mappings
