+++ src/FiapCloundGames.API/Application/Mappings/BibliotecaProfile.cs
@@ -2,7 +2,7 @@
 using FiapCloundGames.API.Application.Dtos.Biblioteca;
 using FiapCloundGames.API.Domain.Entities;
 
-namespace FiapCloundGames.API.Configuration.Mappings
+namespace FiapCloundGames.API.Application.Mappings
 {
     public class BibliotecaProfile : Profile
     {
