+++ src/FiapCloundGames.API/Application/Mappings/JogoProfile.cs
@@ -3,7 +3,7 @@ using FiapCloundGames.API.Application.Dtos.Jogos;
 using FiapCloundGames.API.Application.Dtos.Promocao;
 using FiapCloundGames.API.Domain.Entities;
 
-namespace FiapCloundGames.API.Configuration.Mapping
+namespace FiapCloundGames.API.Application.Mappings
 {
     public class JogoProfile : Profile
     {
